move tests to root
