+++ src/__tests__/Command.test.ts
@@ -1,8 +1,8 @@
 import { ActionImpl } from "..";
-import { Action } from "../../types";
-import { createAction } from "../../utils";
-import { Command } from "../Command";
-import { history } from "../HistoryImpl";
+import { Action } from "../types";
+import { createAction } from "../utils";
+import { Command } from "../action/Command";
+import { history } from "../action/HistoryImpl";
 
 const negate = jest.fn();
 const perform = jest.fn().mockReturnValue(negate);
